requirements.txt
asgiref==3.2.7
Django==3.0.5
django-leaflet==0.26.0
djangorestframework==3.11.0
djangorestframework-gis==0.15
gunicorn==20.0.4
psycopg2-binary==2.8.5
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
